Feature paperbin in some modules and DTOs classes

+++ parking/Controllers/ClientController.cs
@@ -4,51 +4,68 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Xml.XPath;
+using parking.DTO;
 using parking.Models;
 
 namespace parking.Controllers
 {
-    internal class ClientController
+    internal class ClientController: DataBaseController
     {
         private Helpers.Helpers h;
-        public ClientController() {
+        public ClientController()
+        {
             h = new Helpers.Helpers();
         }
 
-        public IEnumerable<dynamic> getClients(string searchFilter)
+        public IEnumerable<ClientDTO> getClients(string searchFilter = "", bool isDel = false)
         {
-            IEnumerable<dynamic> clients = new List<CLIENTS>();
             try
             {
                 using (PARKINGEntities db = new PARKINGEntities())
                 {
-                    clients = (from c in db.CLIENTS join u
-                               in db.USERS on c.USER_ID equals u.USER_CODE
-                               where (string.IsNullOrEmpty(searchFilter) ? (c.CLIENT_NAME + " "+ c.CLIENT_LASTNAME).Contains(searchFilter) && c.IS_DEL==false : c.IS_DEL==false)
-                               select new
-                               {
-                                   c.CLIENT_CODE,
-                                   c.CLIENT_NAME,
-                                   c.CLIENT_LASTNAME,
-                                   c.CLIENT_EMAIL,
-                                   c.CLIENT_ADDRESS,
-                                   c.CLIENT_PHONE,
-                                   c.USER_ID,
-                                   c.IS_DEL,
-                                   u.USER_NAME,
-                                   c.INSERTED_AT,
-                                   c.CLIENT_DNI
-                               }).ToList();
+                    var query = from c in db.CLIENTS
+                                join u in db.USERS on c.USER_ID equals u.USER_CODE
+                                where c.IS_DEL == isDel
+                                select new ClientDTO
+                                {
+                                    CLIENT_CODE = c.CLIENT_CODE,
+                                    CLIENT_NAME = c.CLIENT_NAME,
+                                    CLIENT_LASTNAME = c.CLIENT_LASTNAME,
+                                    CLIENT_EMAIL = c.CLIENT_EMAIL,
+                                    CLIENT_ADDRESS = c.CLIENT_ADDRESS,
+                                    CLIENT_PHONE = c.CLIENT_PHONE,
+                                    USER_ID = c.USER_ID,
+                                    IS_DEL = c.IS_DEL,
+                                    USER_NAME = u.USER_NAME,
+                                    INSERTED_AT = c.INSERTED_AT,
+                                    CLIENT_DNI = c.CLIENT_DNI
+                                };
+
+                    if (!string.IsNullOrEmpty(searchFilter))
+                    {
+                        query = query.Where(c =>
+                            c.CLIENT_CODE.Contains(searchFilter) ||
+                            (c.CLIENT_NAME + " " + c.CLIENT_LASTNAME).Contains(searchFilter) ||
+                            c.CLIENT_PHONE.Contains(searchFilter) ||
+                            c.CLIENT_ADDRESS.Contains(searchFilter) ||
+                            c.CLIENT_DNI.Contains(searchFilter) ||
+                            c.INSERTED_AT.ToString().Contains(searchFilter));
+                    }
+
+                    return query.OrderBy(c => c.CLIENT_CODE).ToList();
                 }
             }
             catch (Exception ex)
             {
-                h.MsgError(ex.ToString());
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
-            return clients;
+
+            return Enumerable.Empty<ClientDTO>();
         }
 
-        public CLIENTS getClient(string clientId){
+
+        public CLIENTS getClient(string clientId)
+        {
             CLIENTS client = new CLIENTS();
             try
             {
@@ -59,7 +76,7 @@ namespace parking.Controllers
             }
             catch (Exception ex)
             {
-                h.MsgError(ex.ToString());
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
             return client;
 
@@ -70,16 +87,17 @@ namespace parking.Controllers
             int result = 0;
             try
             {
-                using (PARKINGEntities db= new PARKINGEntities())
+                using (PARKINGEntities db = new PARKINGEntities())
                 {
                     db.CLIENTS.Add(client);
                     result = db.SaveChanges();
                 }
 
 
-            }catch(Exception ex)
+            }
+            catch (Exception ex)
             {
-                h.MsgError(ex.ToString());
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
 
             return result;
@@ -92,13 +110,13 @@ namespace parking.Controllers
             {
                 using (PARKINGEntities db = new PARKINGEntities())
                 {
-                   db.Entry(client).State = System.Data.Entity.EntityState.Modified;
-                   result = db.SaveChanges();
+                    db.Entry(client).State = System.Data.Entity.EntityState.Modified;
+                    result = db.SaveChanges();
                 }
             }
             catch (Exception ex)
             {
-                h.MsgError(ex.ToString());
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
             return result;
         }
@@ -110,6 +128,11 @@ namespace parking.Controllers
             {
                 using (PARKINGEntities db = new PARKINGEntities())
                 {
+                    if(HasReferences(db,db.CHECK_IN,c => c.CLIENT_DNI == clientId))
+                    {
+                        h.MsgError(Helpers.App.Msg0019);
+                        return 0;
+                    }
                     CLIENTS client = db.CLIENTS.Find(clientId);
                     db.CLIENTS.Attach(client);
                     db.CLIENTS.Remove(client);
@@ -118,7 +141,7 @@ namespace parking.Controllers
             }
             catch (Exception ex)
             {
-                h.MsgError(ex.ToString());
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
             return result;
