+++ parking/Controllers/JobPositionController.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data.Entity.Infrastructure;
+using System.Data.SqlClient;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -7,7 +9,7 @@ using parking.Models;
 
 namespace parking.Controllers
 {
-    internal class JobPositionController
+    internal class JobPositionController: DataBaseController
     {
         private JOB_POSITIONS jobPosition;
         Helpers.Helpers h = new Helpers.Helpers();
@@ -15,23 +17,20 @@ namespace parking.Controllers
             jobPosition = new JOB_POSITIONS();
         }
 
-        public List<JOB_POSITIONS> getJobPositions(string searchFilter)
+        public List<JOB_POSITIONS> getJobPositions(string searchFilter,bool isDel)
         {
             List<JOB_POSITIONS> jobPositions = new List<JOB_POSITIONS>();
             try
             {
                 using (PARKINGEntities db = new PARKINGEntities())
                 {
-                    if (searchFilter != "")
-                    {
-                        jobPositions = db.JOB_POSITIONS.Where(jp => jp.IS_DEL==false && jp.DESCRIPTION_JOB_POSITION.Contains(searchFilter)).ToList();
-                        
-                    }
-                    else
-                    {
-                        jobPositions = db.JOB_POSITIONS.Where(jp => jp.IS_DEL == false).ToList();
-                    }
+                   jobPositions= db.JOB_POSITIONS.Where(j => !String.IsNullOrEmpty(searchFilter) ? (j.JOB_POSITION_CODE.Contains(searchFilter) || j.DESCRIPTION_JOB_POSITION.Contains(searchFilter) || j.INSERTED_AT.ToString().Contains(searchFilter)) && j.IS_DEL == isDel : j.IS_DEL == isDel).ToList();
                 }
+            }
+            catch (SqlException ex)
+            {
+                h.MsgError("ERROR EN LA BASE DE DATOS: " + ex.Message.ToUpper());
+
             }
             catch (Exception ex)
             {
@@ -49,6 +48,11 @@ namespace parking.Controllers
                 {
                     jps = db.JOB_POSITIONS.Find(id);
                 }
+            }
+            catch (SqlException ex)
+            {
+                h.MsgError("ERROR EN LA BASE DE DATOS: " + ex.Message.ToUpper());
+
             }
             catch (Exception ex)
             {
@@ -68,6 +72,11 @@ namespace parking.Controllers
                     result= db.SaveChanges();
                 }
             }
+            catch (SqlException ex)
+            {
+                h.MsgError("ERROR EN LA BASE DE DATOS: " + ex.Message.ToUpper());
+
+            }
             catch (Exception ex)
             {
                 h.MsgError(ex.Message);
@@ -86,6 +95,11 @@ namespace parking.Controllers
                     result = db.SaveChanges();
                 }
             }
+            catch (SqlException ex)
+            {
+                h.MsgError("ERROR EN LA BASE DE DATOS: " + ex.Message.ToUpper());
+
+            }
             catch (Exception ex)
             {
                 h.MsgError(ex.Message);
@@ -100,14 +114,25 @@ namespace parking.Controllers
             {
                 using (PARKINGEntities db = new PARKINGEntities())
                 {
+                    if (HasReferences(db, db.EMPLOYEES, e => e.JOB_POSITION_CODE == jps.JOB_POSITION_CODE))
+                    {
+                        h.MsgError(Helpers.App.Msg0019);
+                        return 0;
+                    }
+
                     db.JOB_POSITIONS.Attach(jps);
                     db.JOB_POSITIONS.Remove(jps);
                     result = db.SaveChanges();
                 }
             }
+            catch (SqlException ex)
+            {
+                h.MsgError("ERROR EN LA BASE DE DATOS: " + ex.Message.ToUpper());
+
+            }
             catch (Exception ex)
             {
-                h.MsgError(ex.Message);
+                h.MsgError("ERROR INESPERADO: " + ex.Message.ToUpper());
             }
             return result;
         }
